Downgrade to Java 8

+++ src/test/java/clipper2/BenchmarkClipper2.java
@@ -66,8 +66,8 @@ public class BenchmarkClipper2 {
 	}
 
 	private static Point64 MakeRandomPt(int maxWidth, int maxHeight, Random rand) {
-		long x = rand.nextLong(maxWidth);
-		long y = rand.nextLong(maxHeight);
+		long x = rand.nextInt(maxWidth);
+		long y = rand.nextInt(maxHeight);
 		return new Point64(x, y);
 	}
